Found a bug where the robot could go 1 step beyond the table X limit
Given that we flipped the grid such that the negative portion of the X axis represents what the user thinks is the positive portion (so that the origin could be in the SW corner as stated by the problem) a bit of the logic for handling the move limit when going eastward was overlooked.

Our next task is to add proper unit tests to try and catch scenarios like this. Yes after the fact. Judge all you like.